sui-graphql-client: fix transaction(s) query (#32)
The TransactionBlock bcs in GraphQL now encodes the TransactionData. This PR fixes the logic to correctly deserialize the transaction bcs data and signatures into a SignedTransaction.

As the service does not serve yet the new data, I tested it locally by starting a local network and GraphQL server. For now, I put an #ignore on the test.